requirements.txt
streamlit>=1.30.0
pandas>=2.2.0
openai>=1.8.0
chromadb>=0.4.22
python-dotenv>=1.0.0